WooCommerce: Make api-plan general purpose
This takes the api-plan code that was originally developed under
state/ui/product and places it under state/api-plan. It also seeks to
make more general use of the code and adds a README to explain its
operation.

This is really just the reducer/action code, there are still selectors
to be written and more helper functions to be ported/written.

Woo: Change API Plan to Action List

This changes the name of the API Plan to Action List because "Plan"
is an overloaded term within Calypso.

Woo Action-List: Post-rebase cleanup

This removes some files that lingered around after the rebase,
and make a few small adjustments to code.

Split action list end into success/failure

This splits the action for ending a step into two separate actions for
success and failure, respectively. This will line up with the data layer
and the success/failure actions for request handling.

Set Date.now() from the reducer.

This changes the action-list code for each step to set the time to
Date.now() in the reducer instead of in the action creator. This is due
to the fact that the actions are usually created long before they're
dispatched, so the timing ends up being inaccurate.

Update action-list README

Some things have changed in the system and become more clear.

Woo Action-List: Fix artifact from rebase

This removes the old `apiPlan` reducer variable that got left
over from a rebase operation.